BE_py/book-platform/search-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pymysql
opensearch-py
